requirements.txt
# LangChain & LLM
langchain==1.2.10
langchain-core==1.2.17
langchain-google-genai==4.2.1
google-genai==1.66.0

# LangGraph
langgraph==1.0.10

# Data & Validation
pydantic==2.12.5
pydantic_core==2.41.5

# Research
tavily-python==0.7.22

# Async & HTTP
httpx==0.28.1
aiohttp==3.13.3

# Configuration
python-dotenv==1.0.1

# Observability
mlflow==3.10.1

# Utilities
python-dateutil==2.8.2
